Feat: [#32] 추가 관리자 로그인 기능
로그인, 로그아웃 세션 추가하였습니다.

+++ module-admin/src/main/java/com/adminapplication/dto/AllBoardsResponseDto.java
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AllBoardsResponseDto {
-    private Integer id;
-    private Integer userId;
+    private Long id;
+    private Long userId;
     private String username;
     private String title;
     private String status;
